Refactors ontology documentation structure
Updates the documentation generation process to use "all" instead of "only" for property constraints, and replaces uses of xsd datatypes with owl::Thing where appropriate for greater flexibility.
This change involves updates to class definitions and aims to improve the consistency and maintainability of the generated documentation.
A new Law doc and BuildingUnit doc are created.